all: rename packages
- Rename cmd/runlocal to cmd/tailsql.
  For now, the functionality has not changed.

- Move ./tailsql to ./server/tailsql.
  This lets us build a "tailsql" binary at root.
