+++ server/tailsql/options.go
