+++ server/tailsql/static/script.js
