+++ server/tailsql/ui.tmpl
