+++ server/tailsql/utils.go
